// frontend/src/domain/store/store.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type StoreSearchVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    code?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    nameCode?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    siteId?: {
    equalTo?: number | null;
    equalAny?: Array<number> | null;
    notEqualTo?: number | null;
    equalAnyOrNull?: Array<number> | null;
    notEqualAll?: Array<number> | null;
  } | null;
    codeOrName?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
  } | null;
  page?: {
    first?: number | null;
    offset?: number | null;
  } | null;
};

export type StoreSearchResult = {
  stores: ({
  __typename: "StoreConnector";
} & {
  totalCount: number;
  nodes: Array<{
  __typename: "StoreNode";
  id: string;
  code: string;
  storeName: string;
  siteId: number;
}>;
});
};

export const StoreSearch = {
  query: "query storeSearch($filter: StoreFilterInput, $page: PaginationInput) {\n  stores(filter: $filter, sort: [{key: name}], page: $page) {\n    ... on StoreConnector {\n      __typename\n      totalCount\n      nodes {\n        __typename\n        id\n        code\n        storeName\n        siteId\n      }\n    }\n  }\n}",
} as TypedDocument<StoreSearchResult, StoreSearchVariables>;